Add files via upload
I made this project keeping in mind the current existing system. In the present era where time proves to be the most important asset for an individual. So, by replacing the current register system to fully computerised one, will not only save time but also accuracy, reliability and uniformity can be maintained.